cleanup: set long and short date formats in independent branches
The short and long date formats where initialized in a common if-branch.
However, inside the if-branch the code rechecked which of the two should
be initialized. This could make sense if there was some common code between
the two, but there wasn't. Therefore, make this two independent branches
to avoid one nesting-level.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>