profile: correctly display dives with multiple pressure sensors
When removing the MAX_CYLINDERS restriction, the population of the
pressure profile data has been broken. The pressure data is organized
as a nr_cylinders x nr matrix, where nr is the number of samples. However,
an oversized array was allocated and then the nr_entries value set
to the actually needed number of entries. Obviously, this messed up the
matrix accesses.

Do a quick fix by reallocating the array at the end of the function.
We can't simply keep the old nr value, because the profile would attempt
to render too many entries. Probably not a big disaster, because the
entries are empty anyway, but still...

Fixes #2877

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>