Handle zero length segments when replanning
When setting up a dive for replanning, we ignored zero
length segments as those tend to be generated by gas
changes. But it is possible to enter those in the planner
and the replanning should not ignore those. So be
more clever about gas changes. Let's add 10 seconds
so we are not at two depths at the same time and
help since add_stop also does not like zero length
segments (it thinks we are trying to replace a waypoint).

Fixes #2901

Signed-off-by: Robert C. Helling <helling@atdotde.de>