Proof of concept: add C++ iterators to dive_table
This is not intended for merging, therefore I'm not writing a
full commit message. It proves how trivial it is to add C++
style iterators to our table structures.

In this case, there is not much gain (except less polution of
outer scope). However, for cylinders, weights, etc. this might
make the code distinctly nicer.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>