Interpret divelist printing templates
This is a first step of an efford to get rid of the Grantlee
dependency. This implements template processing for those
constructs used in our divelist and statistics printing templates.

It implements a template parser for loops over
dives, cylinders and year and variable replacement. As the
previous Grantlee code, it does not really use Qt's
QObject introspection capabilities but reuses the old
long chain of if-else-statements.

The grantlee code is not yet removed.

Signed-off-by: Robert C. Helling <helling@atdotde.de>