cleanup: when planning a dive, set dive computer to first dc
When planning a dive, dc_number was set to 1, which actually is
the second dc! The code seems to handle this gracefully, however
it appears weird. Let's set dc_number to 0 instead.

Originally, the assignment was introduced in a422957cd6525b9753
and moved later in 4f5621c4c6acc3a.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>