translations: initialize water type strings on demand
The water type strings were global static and therefore passed through
gettextFromC::tr() before main(). One would hope to get a warning
in such a case, but this is not the case.

Therefore, make the array local to a function, which guarantees
that it will be initialized on first invocation in a thread safe
manner. This is in principle the idiomatic C++ "singleton" pattern.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>