Header line: Update INSTALL file to add libbluetooth-dev etc.
When attempting to compile under Debian and Ubuntu the build
script would get stuck looking for bluez. This change adds
libbluetooth-dev to the suggested install packages for Debian and
Ubuntu.

Thus update also changes the yum install command in the Fedora
instructions to the newer dnf command as well as updating the
versions listed of Debiand and Ubuntu to the latest versions.

Signed-off-by: Jason Bramwell <jb2cool@gmail.com>